fix: use only item id and file extension for local file paths
Remote filenames with URL-encoded paths were somehow causing the app
to think the directory for a filename didn't exist. This appeared to happen
for anchor.fm apps, which include a full cloudfront URL as the last path
element in the media URL. Changing this to just use <id>.<extension> locally
seemed to work around the issue.